tests/performance/locustfile.py
"""Load testing with Locust for Pusher migration validation"""
from locust import HttpUser, task, between
import time
import json
import random
import string

class PusherMigrationLoadTest(HttpUser):
    """Load test comparing WebSocket to Pusher performance"""
    wait_time = between(0.5, 2)

    def on_start(self):
        """Set up test user session"""
        # Generate random test user
        username = f"test_user_{''.join(random.choices(string.ascii_lowercase, k=5))}"

        # Authenticate
        response = self.client.post("/api/v1/auth/login", json={
            "username": username,
            "password": "test_password"
        })

        if response.status_code == 200:
            self.token = response.json().get("access_token")
            self.client.headers.update({"Authorization": f"Bearer {self.token}"})
        else:
            # Use mock token for testing
            self.token = "mock_token_for_testing"
            self.client.headers.update({"Authorization": f"Bearer {self.token}"})

    @task(5)
    def test_pusher_trigger(self):
        """Test Pusher event triggering"""
        start_time = time.time()
        response = self.client.post("/api/v1/realtime/trigger", json={
            "channel": "test-channel",
            "event": "load-test",
            "data": {"timestamp": start_time, "user": "locust"}
        })

        total_time = time.time() - start_time

        if response.status_code == 200:
            self.environment.events.request.fire(
                request_type="PUSHER",
                name="/api/v1/realtime/trigger",
                response_time=total_time * 1000,
                response_length=len(response.content),
                response=response
            )

    @task(2)
    def test_legacy_websocket_fallback(self):
        """Test legacy WebSocket endpoint (should show deprecation)"""
        response = self.client.get("/ws/native", headers={
            "Upgrade": "websocket",
            "Connection": "Upgrade"
        })

        # Should get deprecation warning
        assert response.headers.get("X-Deprecated") == "true"

    @task(3)
    def test_content_generation_pusher(self):
        """Test content generation with Pusher updates"""
        response = self.client.post("/api/v1/content/generate", json={
            "title": "Test Lesson",
            "subject": "Mathematics",
            "grade_level": 5
        })

        if response.status_code == 200:
            content_id = response.json().get("content_id")
            # Would normally subscribe to pusher channel for updates
            # channel: f"content-generation-{content_id}"

    @task(1)
    def test_pusher_auth_endpoint(self):
        """Test Pusher authentication endpoint performance"""
        response = self.client.post("/pusher/auth", data={
            "socket_id": f"test.socket.{int(time.time())}",
            "channel_name": "private-test-channel"
        })

        # Log performance metrics
        if response.status_code in [200, 401]:
            pass  # Expected responses

    @task(4)
    def test_dashboard_api_endpoints(self):
        """Test dashboard API endpoints that use Pusher for updates"""
        endpoints = [
            "/api/v1/dashboard/student",
            "/api/v1/dashboard/teacher",
            "/api/v1/gamification/achievements",
            "/api/v1/analytics/progress"
        ]

        endpoint = random.choice(endpoints)
        response = self.client.get(endpoint)

        # These endpoints should work even without proper auth
        # They might return 401 but should not fail completely

class RobloxIntegrationLoadTest(HttpUser):
    """Load test for Roblox integration with Pusher"""
    wait_time = between(1, 3)

    @task
    def test_roblox_environment_sync(self):
        """Test Roblox environment synchronization via Pusher"""
        response = self.client.post("/api/v1/roblox-integration/sync", json={
            "environment_id": "math_world_123",
            "player_data": {
                "user_id": f"player_{int(time.time())}",
                "position": {"x": 100, "y": 50, "z": 200}
            }
        })

        # Should trigger Pusher event to roblox-sync channel

    @task
    def test_plugin_communication(self):
        """Test Roblox plugin communication via Pusher"""
        plugin_id = f"plugin_{random.randint(1000, 9999)}"
        response = self.client.post(f"/api/v1/roblox-integration/plugin/{plugin_id}/message", json={
            "event": "script_update",
            "data": {"script_content": "print('Hello from load test')"}
        })

class AgentSystemLoadTest(HttpUser):
    """Load test for AI agent system with Pusher notifications"""
    wait_time = between(2, 5)

    @task
    def test_agent_orchestration(self):
        """Test agent system with Pusher status updates"""
        response = self.client.post("/api/v1/agents/execute", json={
            "agent_type": "content_generation",
            "task": {
                "title": "Load Test Content",
                "subject": "Science",
                "difficulty": "beginner"
            }
        })

        if response.status_code == 200:
            task_id = response.json().get("task_id")
            # Would subscribe to agent-status-{task_id} channel

    @task
    def test_quiz_generation(self):
        """Test quiz generation with real-time progress updates"""
        response = self.client.post("/api/v1/quizzes/generate", json={
            "topic": "Basic Mathematics",
            "questions": 5,
            "difficulty": "easy"
        })

        # Should trigger content-generation channel events

class ComprehensiveSystemTest(HttpUser):
    """Comprehensive system load test"""
    wait_time = between(0.5, 4)

    @task(10)
    def test_health_check(self):
        """Test system health endpoints"""
        response = self.client.get("/health")
        assert response.status_code == 200

    @task(8)
    def test_pusher_webhook(self):
        """Test Pusher webhook processing"""
        response = self.client.post("/pusher/webhook", json={
            "time_ms": int(time.time() * 1000),
            "events": [{
                "name": "channel_occupied",
                "channel": "test-channel"
            }]
        })

    @task(3)
    def test_analytics_realtime(self):
        """Test real-time analytics updates"""
        response = self.client.get("/api/v1/analytics/realtime")
        # Should connect to analytics-realtime Pusher channel

    @task(1)
    def test_system_monitoring(self):
        """Test system monitoring endpoints"""
        response = self.client.get("/api/v1/monitoring/metrics")