requirements.txt
# Core Web Frameworks
fastapi==0.116.1
uvicorn[standard]==0.35.0
flask==3.0.0
Flask-Cors==6.0.1
asgiref>=3.8.1
greenlet>=3.2.4

# Phase 1 Dependencies (OpenAI GPT-5 & OAuth 2.1)
openai>=1.25.0  # Updated for GPT-5 support
authlib>=1.3.0  # OAuth 2.1 compliance
cryptography>=42.0.0  # Enhanced security

# LangChain Ecosystem (2025 Latest Versions - Pydantic v2 Compatible)
langchain>=0.3.20,<0.4.0
langchain-core>=0.3.40,<0.4.0
langchain-community>=0.3.20,<0.4.0
langchain-openai>=0.2.10,<0.3.0
langchain-anthropic>=0.3.0,<0.4.0
langgraph>=0.2.65,<0.3.0
langgraph-checkpoint-sqlite>=2.0.0,<3.0.0
langsmith>=0.1.0,<0.2.0
tiktoken==0.11.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.30.0
redis[hiredis]>=5.0.8  # Latest Redis client with C speedups
redis-om>=0.2.0  # Redis object mapping
aiosqlite==0.20.0
psycopg2-binary==2.9.9
supabase>=2.0.0,<3.0.0  # Supabase Python client
langcache>=0.1.0  # Semantic caching for AI responses

# Authentication & Security
PyJWT==2.10.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
requests-oauthlib==2.0.0

# Data Validation (Compatible with LangChain 0.3.x)
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv==1.0.0
pyyaml==6.0.1

# WebSockets & Realtime
websockets==15.0.1
websocket-client==1.8.0
python-socketio>=5.10,<6
flask-socketio==5.3.6
pusher==3.3.2
aiohttp==3.12.14
aiofiles==24.1.0

# HTTP Clients
httpx==0.28.1
requests==2.32.5

# Monitoring & Logging
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.0.0
structlog==25.4.0
colorama==0.4.6
sentry-sdk[fastapi,pure_eval,sqlalchemy]==2.37.1

# OpenTelemetry (Observability)
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation>=0.42b0,<1.0.0
opentelemetry-instrumentation-fastapi>=0.42b0,<1.0.0
opentelemetry-instrumentation-httpx>=0.42b0,<1.0.0
opentelemetry-instrumentation-sqlalchemy>=0.42b0,<1.0.0
opentelemetry-instrumentation-redis>=0.42b0,<1.0.0
opentelemetry-instrumentation-psycopg2>=0.42b0,<1.0.0
opentelemetry-instrumentation-logging>=0.42b0,<1.0.0
# Using only OTLP HTTP exporter to avoid gRPC conflicts
opentelemetry-exporter-otlp-proto-http>=1.21.0,<2.0.0
opentelemetry-propagator-b3>=1.21.0,<2.0.0

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==6.2.1
pytest-mock==3.12.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
pytest-html==4.1.1
pytest-json-report==1.5.0
safety>=3.0.0  # Security vulnerability scanning

# Development Tools
debugpy==1.8.0
# pdbpp==0.10.3  # Temporarily disabled due to fancycompleter incompatibility
ipdb==0.13.13
memory-profiler==0.61.0
line-profiler==4.1.2
py-spy==0.3.14

# Code Quality
black==25.1.0
flake8==7.3.0
mypy==1.7.1
ipython~=8.26

# System Monitoring
psutil>=5.9,<6
py-cpuinfo==9.0.0
python-json-logger==2.0.7

# AI/ML Libraries
numpy==1.26.4
scipy>=1.11.0,<1.12.0
scikit-learn>=1.5.0,<2.0.0
sentence-transformers==5.1.0
faiss-cpu==1.8.0

# External Data Sources
# ddgs>=1.0.0  # REMOVED: Not imported or used anywhere

# Utilities
python-dateutil==2.9.0.post0
pandas==2.2.2
pytz>=2024.1,<2026.0
tenacity==9.1.2

# Deployment
gunicorn==23.0.0
supervisor==4.2.5

# Documentation
mkdocs==1.6.0
mkdocs-material==9.6.20

# Security
urllib3==2.5.0
cryptography>=42.0.0  # Replaces vulnerable ecdsa

# GraphQL Dependencies
ariadne==0.23.0                # GraphQL server with ASGI support built-in
ariadne-codegen==0.13.0        # Python code generation from schema
aiodataloader==0.4.0           # DataLoader for N+1 query prevention
graphql-core>=3.2.0            # GraphQL core implementation

# Payments
stripe>=7.0.0,<11.0.0          # Stripe Python SDK for webhooks

# Celery - Distributed Task Queue (Latest stable as of 2025)
# Reference: https://docs.celeryq.dev/en/stable/
celery[redis]==5.4.0           # Latest stable with Redis support
flower==2.0.1                  # Real-time monitoring for Celery
kombu>=5.3.0                   # Messaging library for Celery
vine>=5.1.0                    # Promise library for Celery
billiard>=4.2.0                # Python multiprocessing fork
amqp>=5.2.0                    # AMQP client library

# Email Services (Render/Supabase Deployment)
sendgrid==6.12.5               # SendGrid email service (latest stable)
premailer>=3.10.0              # CSS inlining for emails
bleach>=6.1.0                  # HTML sanitization
beautifulsoup4>=4.12.0         # HTML parsing
markdown>=3.6                  # Markdown to HTML conversion
jinja2>=3.1.4                  # Template engine
python-http-client>=3.3.7      # SendGrid dependency
